requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
feedparser==6.0.10
idna==3.4
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.2
sgmllib3k==1.0.0
six==1.16.0
urllib3==1.26.14
